test(reporter-tippy): move to uvu
