+++ packages/reporter-tippy/package.json
@@ -25,8 +25,8 @@
     "build": "pnpm prebuild && cross-env NODE_ENV=production rollup -c",
     "dev": "rollup -cw",
     "prepublishOnly": "pnpm build && pnpm test",
-    "test": "jest",
-    "test:ci": "jest --ci --coverage"
+    "test": "uvu -r tsm -r global-jsdom/register tests -i common",
+    "test:ci": "nyc -n src pnpm test"
   },
   "dependencies": {
     "@felte/common": "workspace:*"
@@ -35,6 +35,7 @@
     "tippy.js": "^6.0.0"
   },
   "devDependencies": {
+    "@felte/core": "workspace:*",
     "felte": "workspace:*",
     "svelte": "^3.46.4"
   },
